Add governance policies
Add a number of policies and documents about CDEvents governance
and processes. A lot of content is inspired by the Tekton community
governance documents and adapted for CDEvents.

Key new things added:
- governing boards elections (eligibility, etc)
- contributor roles and ladder (eligibility, etc)

Some new mailing lists will have to be created.

The current bootstrap committee includes 5 individuals.
My proposal is to keep a team of five. For continuity, I propose that
- We held elections for three of the seats, to ensure continuity in the
  project governance.
- Emil and Andrea have been acting as project governance since the
  beginning, so their seats that be maintained until after 1 year
  after the election
- Nattias, Erik and Steve seats are up for election

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>