Add governance policies (#38)
* Add governance policies

Add a number of policies and documents about CDEvents governance
and processes. A lot of content is inspired by the Tekton community
governance documents and adapted for CDEvents.

Key new things added:
- governing boards elections (eligibility, etc)
- contributor roles and ladder (eligibility, etc)

Some new mailing lists will have to be created.

The current bootstrap committee includes 5 individuals.
My proposal is to keep a team of five. For continuity, I propose that
- We held elections for three of the seats, to ensure continuity in the
  project governance.
- Emil and Andrea have been acting as project governance since the
  beginning, so their seats that be maintained until after 1 year
  after the election
- Mattias, Erik and Steve seats are up for election

* Add back the cdevents-code-of-conduct mailing list

This will have to be updated on the code of conduct in the .github repo
as well as in the community page on the website

* Align on governing board naming

* Remove the reviewer role

Remove the reviewer role as agreed during the WG meeting on 23.11
since the role is not used today and we'd like to keep policies
as simple as possible.

Clarify the stepping down / emeritus process.

---------

Signed-off-by: Andrea Frittoli <andrea.frittoli@gmail.com>